+++ material.py
@@ -3,7 +3,7 @@ import taichi as ti
 # Here we use neo-Hookean model
 
 @ti.func
-def DeterminantGrad(A: ti.types.matrix(3, 3, ti.f32)) -> ti.types.matrix(3, 3, ti.f32):
+def DeterminantGrad(A: ti.types.matrix(3, 3, ti.f64)) -> ti.types.matrix(3, 3, ti.f64):
     dJdA = ti.Matrix([[0.0 for i in range(3)] for j in range(3)])
     dJdA[0, 0] = A[1, 1] * A[2, 2] - A[1, 2] * A[2, 1]
     dJdA[0, 1] = A[2, 0] * A[1, 2] - A[1, 0] * A[2, 2]
@@ -17,11 +17,11 @@ def DeterminantGrad(A: ti.types.matrix(3, 3, ti.f32)) -> ti.types.matrix(3, 3, t
     return dJdA    
 
 materialTypeDict = dict(
-    density = ti.f32,
-    youngs_modulus = ti.f32,
-    possions_ratio = ti.f32,
-    lam = ti.f32,
-    mu = ti.f32,
+    density = ti.f64,
+    youngs_modulus = ti.f64,
+    possions_ratio = ti.f64,
+    lam = ti.f64,
+    mu = ti.f64,
 )
 
 @ti.data_oriented
@@ -34,7 +34,7 @@ class Material:
         self.mu = 0.0
 
     @ti.kernel
-    def Initialize(self, density: ti.f32, youngs_modulus: ti.f32, poissons_ratio: ti.f32):
+    def Initialize(self, density: ti.f64, youngs_modulus: ti.f64, poissons_ratio: ti.f64):
         self.density = density
         self.youngs_modulus = youngs_modulus
         self.poissons_ratio = poissons_ratio
@@ -42,7 +42,7 @@ class Material:
         self.mu = youngs_modulus / (2 * (1 + poissons_ratio))
 
     @ti.kernel
-    def ComputeEnergyDensity(self, F: ti.types.matrix(3, 3, ti.f32)) -> ti.f32:
+    def ComputeEnergyDensity(self, F: ti.types.matrix(3, 3, ti.f64)) -> ti.f64:
         C = F.transpose() @ F
         J = F.determinant()
         Ic = C[0, 0] + C[1, 1] + C[2, 2]
@@ -54,7 +54,7 @@ class Material:
         return mu / 2 * (Ic - dim) + la / 2 * (J - alpha) * (J - alpha) - 0.5 * mu * ti.log(Ic + delta)
 
     @ti.kernel
-    def ComputeStressDensity(self, F: ti.types.matrix(3, 3, ti.f32)) -> ti.types.matrix(3, 3, ti.f32):
+    def ComputeStressDensity(self, F: ti.types.matrix(3, 3, ti.f64)) -> ti.types.matrix(3, 3, ti.f64):
         C = F.transpose() @ F
         J = F.determinant()
         Ic = C[0, 0] + C[1, 1] + C[2, 2]
@@ -67,7 +67,7 @@ class Material:
         return (1 - 1 / (Ic + delta)) * mu * F + la * (J - alpha) * dJdF
 
 @ti.func
-def ComputeEnergyDensity(F:ti.types.matrix(3,3,ti.f32),lam:ti.f32, mu:ti.f32):
+def ComputeEnergyDensity(F:ti.types.matrix(3,3,ti.f64),lam:ti.f64, mu:ti.f64):
     C = F.transpose() @ F
     J = F.determinant()
     Ic = C[0, 0] + C[1, 1] + C[2, 2]
@@ -77,7 +77,7 @@ def ComputeEnergyDensity(F:ti.types.matrix(3,3,ti.f32),lam:ti.f32, mu:ti.f32):
     return mu / 2 * (Ic - dim) + lam / 2 * (J - alpha) * (J - alpha) - 0.5 * mu * ti.log(Ic + delta)
 
 @ti.func
-def ComputeStressDensity(F:ti.types.matrix(3,3,ti.f32),lam:ti.f32, mu:ti.f32):
+def ComputeStressDensity(F:ti.types.matrix(3,3,ti.f64),lam:ti.f64, mu:ti.f64):
     C = F.transpose() @ F
     J = F.determinant()
     Ic = C[0, 0] + C[1, 1] + C[2, 2]
